% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Frame.R
\name{cbind.R6Frame}
\alias{cbind.R6Frame}
\title{R6Frame: cbind}
\usage{
cbind.R6Frame(x, ...)
}
\arguments{
\item{x}{A \code{R6Frame}.}

\item{...}{Additional parameters passed to \code{cbind}.}
}
\description{
\code{\link[base]{cbind}} method for \code{R6Frame} objects. Labels and associations
are also merged if you are binding other R6Frames.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

