% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frame_df.R
\name{cbind}
\alias{cbind}
\title{R6Frame: cbind}
\usage{
cbind(x, ...)
}
\arguments{
\item{x}{A \code{R6Frame}.}

\item{...}{Additional parameters passed to \code{cbind}.}
}
\description{
A new generic for \code{\link[base]{cbind}} which includes a default method, and
a method for \code{R6Frame}. For all other objects, it defaults to the base method.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

