% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frame_dt.R
\name{dcast}
\alias{dcast}
\title{data.table: dcast}
\usage{
dcast(data, ...)
}
\arguments{
\item{data}{A \code{R6Frame} or \code{data.frame}.}

\item{...}{Additional parameters passed to \code{dcast}.}
}
\description{
Same as the \code{data.table} function \code{dcast}. This is a new generic,
because the \code{data.table} version calls \code{reshape2} unless the input
is a \code{data.table}.
}
\author{
Kristian D. Olsen
}
\seealso{
\code{\link[data.table]{dcast}} in \code{data.table} for more information and examples.
}

