% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frame_tbl.R
\name{dplyr_binds}
\alias{bind_cols}
\alias{bind_cols.R6Frame}
\alias{bind_rows}
\alias{bind_rows.R6Frame}
\alias{dplyr_binds}
\title{dplyr: Methods for R6Frame.}
\usage{
bind_rows(x, ...)

\method{bind_rows}{R6Frame}(x, ...)

bind_cols(x, ...)

\method{bind_cols}{R6Frame}(x, ...)
}
\arguments{
\item{x}{A R6Frame object.}

\item{...}{Further arguments passed to \code{dplyr}.}
}
\description{
\code{dplyr} bind methods for R6Frame.
}
\details{
NOTE: In order for these verbs to function properly, dplyr must be attached
with \code{library} or \code{require} before \code{reporttoolDT}. (I think).
}
\author{
Kristian D. Olsen
}

