% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frame_tbl.R
\name{dplyr_joins}
\alias{anti_join.R6Frame}
\alias{dplyr_joins}
\alias{full_join.R6Frame}
\alias{left_join.R6Frame}
\alias{right_join.R6Frame}
\alias{semi_join.R6Frame}
\title{dplyr: Methods for R6Frame.}
\usage{
left_join.R6Frame(x, y, ...)

right_join.R6Frame(x, y, ...)

full_join.R6Frame(x, y, ...)

semi_join.R6Frame(x, y, ...)

anti_join.R6Frame(x, y, ...)
}
\arguments{
\item{x}{A \code{R6Frame} object.}

\item{y}{A \code{R6Frame}, regular \code{data.frame}, \code{data.table} or \code{tbl}.}

\item{...}{Further arguments passed to \code{dplyr}.}
}
\description{
\code{dplyr} join methods for R6Frame.
}
\details{
NOTE: In order for these verbs to function properly, dplyr must be attached
with \code{library} or \code{require} before \code{reporttoolDT}. (I think).
}
\author{
Kristian D. Olsen
}

