% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frame_tbl.R
\name{dplyr_verbs}
\alias{arrange_.R6Frame}
\alias{dplyr_verbs}
\alias{filter_.R6Frame}
\alias{group_by_.R6Frame}
\alias{groups.R6Frame}
\alias{mutate_.R6Frame}
\alias{rename_.R6Frame}
\alias{select_.R6Frame}
\alias{summarise_.R6Frame}
\alias{tbl_vars.R6Frame}
\alias{ungroup.R6Frame}
\title{dplyr: Methods for R6Frame.}
\usage{
tbl_vars.R6Frame(x)

select_.R6Frame(x, ...)

rename_.R6Frame(x, ...)

filter_.R6Frame(x, ...)

arrange_.R6Frame(x, ...)

group_by_.R6Frame(x, ...)

groups.R6Frame(x, ...)

ungroup.R6Frame(x, ...)

mutate_.R6Frame(x, ...)

summarise_.R6Frame(x, ...)
}
\arguments{
\item{x}{A R6Frame object.}

\item{...}{Further arguments passed to \code{dplyr}.}
}
\description{
R6Frames include support for \code{dplyr}. See \code{dplyr} documentation
for further information.
}
\details{
NOTE: In order for these verbs to function properly, dplyr must be attached
with \code{library} or \code{require} before \code{reporttoolDT}. (I think).
}
\author{
Kristian D. Olsen
}

