% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frame_df.R
\name{rbind}
\alias{rbind}
\title{R6Frame: rbind}
\usage{
rbind(x, ...)
}
\arguments{
\item{x}{A \code{R6Frame}.}

\item{...}{Additional parameters passed to \code{rbind}.}
}
\description{
A new generic for \code{\link[base]{rbind}} which includes a default method, and
a method for \code{R6Frame}. For all other objects, it defaults to the base method.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

