% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frame_tidyr.R
\name{tidyr_verbs}
\alias{complete_.R6Frame}
\alias{expand_.R6Frame}
\alias{fill_.R6Frame}
\alias{gather_.R6Frame}
\alias{spread_.R6Frame}
\alias{tidyr_verbs}
\title{tidyr: Methods for R6Frame.}
\usage{
gather_.R6Frame(x, ...)

spread_.R6Frame(x, ...)

complete_.R6Frame(x, ...)

expand_.R6Frame(x, ...)

fill_.R6Frame(x, ...)
}
\arguments{
\item{x}{A \code{R6Frame} object.}

\item{...}{Further arguments passed to \code{tidyr}.}
}
\description{
R6Frames include support for \code{tidyr}. See \code{tidyr} documentation
for further information.
}
\details{
NOTE: In order for these verbs to function properly, tidyr must be attached
with \code{library} or \code{require} before \code{reporttoolDT}. (I think).
}
\author{
Kristian D. Olsen
}

