% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frame_dt.R
\name{melt}
\alias{melt}
\title{data.table: melt}
\usage{
melt(data, ...)
}
\arguments{
\item{data}{A \code{R6Frame} or \code{data.frame}.}

\item{...}{Additional parameters passed to \code{melt}.}
}
\description{
Same as the \code{data.table} function \code{melt}. This is a new generic,
because the \code{data.table} version calls \code{reshape2} unless the input
is a \code{data.table}.
}
\author{
Kristian D. Olsen
}
\seealso{
\code{\link[data.table]{melt}} in \code{data.table} for more information and examples.
}

